+++ pcap4j-core/src/test/java/org/pcap4j/core/PcapsTest.java
@@ -21,7 +21,7 @@ import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.net.InetAddress;
-import java.sql.Timestamp;
+import java.time.Instant;
 import java.util.List;
 
 import static org.junit.Assert.assertEquals;
@@ -103,8 +103,8 @@ public class PcapsTest {
           TimestampPrecision.NANO
         );
     PcapPacket packet = phNano.getNextPacket();
-    assertEquals(1434220771517L, packet.getTimestamp().getTime());
-    assertEquals(517995677, packet.getTimestamp().getNanos());
+    assertEquals(1434220771L, packet.getTimestamp().getEpochSecond());
+    assertEquals(517995677, packet.getTimestamp().getNano());
     phNano.close();
 
     PcapHandle phMicro
@@ -114,10 +114,10 @@ public class PcapsTest {
           TimestampPrecision.MICRO
         );
     packet = phMicro.getNextPacket();
-    System.out.println(packet.getTimestamp().getTime());
-    System.out.println(packet.getTimestamp().getNanos());
-    assertEquals(1434220771517L, packet.getTimestamp().getTime());
-    assertEquals(517995000, packet.getTimestamp().getNanos());
+    System.out.println(packet.getTimestamp().getEpochSecond());
+    System.out.println(packet.getTimestamp().getNano());
+    assertEquals(1434220771L, packet.getTimestamp().getEpochSecond());
+    assertEquals(517995000, packet.getTimestamp().getNano());
     phMicro.close();
   }
 
@@ -162,8 +162,7 @@ public class PcapsTest {
           .payloadBuilder(ab)
           .paddingAtBuild(true);
     Packet packet = eb.build();
-    Timestamp ts = new Timestamp(1234567890123L);
-    ts.setNanos(123456789);
+    Instant ts = Instant.ofEpochSecond(1234567890L,123456789L);
 
     PcapHandle phdMicro = Pcaps.openDead(
       DataLinkType.EN10MB,
@@ -179,8 +178,8 @@ public class PcapsTest {
     PcapHandle phNano1
       = Pcaps.openOffline(tmpFile1,TimestampPrecision.NANO);
     PcapPacket ppacket = phNano1.getNextPacket();
-    assertEquals(1234567890123L, ppacket.getTimestamp().getTime());
-    assertEquals(123456000, ppacket.getTimestamp().getNanos());
+    assertEquals(1234567890L, ppacket.getTimestamp().getEpochSecond());
+    assertEquals(123456000, ppacket.getTimestamp().getNano());
     phNano1.close();
 
     PcapHandle phdNano = Pcaps.openDead(
@@ -197,8 +196,8 @@ public class PcapsTest {
     PcapHandle phNano2
       = Pcaps.openOffline(tmpFile2,TimestampPrecision.NANO);
     ppacket = phNano2.getNextPacket();
-    assertEquals(1234567890123L, ppacket.getTimestamp().getTime());
-    assertEquals(123456789, ppacket.getTimestamp().getNanos());
+    assertEquals(1234567890L, ppacket.getTimestamp().getEpochSecond());
+    assertEquals(123456789, ppacket.getTimestamp().getNano());
     phNano2.close();
   }