+++ pcap4j-packettest/src/test/java/org/pcap4j/packet/IpV4PacketTest.java
@@ -27,7 +27,7 @@ import java.io.StringReader;
 import java.net.Inet4Address;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
-import java.sql.Timestamp;
+import java.time.Instant;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
@@ -410,7 +410,7 @@ public class IpV4PacketTest extends AbstractPacketTest {
 
     PcapHandle handle = Pcaps.openDead(DataLinkType.EN10MB, 65536);
     PcapDumper dumper = handle.dumpOpen(dumpFile);
-    Timestamp ts = new Timestamp(0);
+    Instant ts = Instant.ofEpochSecond(0);
     dumper.dump(ep1, ts);
     dumper.dump(ep2, ts);
     dumper.close();