+++ CHANGES.md
@@ -30,6 +30,7 @@ Future (2.0.0)
         * PcapHandle#loop with RawPacketListener
         * PcapHandle#dispatch with RawPacketListener
         * RawPacketListener
+* Use java.time.Instant for timestamps instead of java.sql.Timestamp.
 
 ### Bug Fixes ###